examples/baby-names/backend/requirements.txt
Flask==3.1.0
flask-cors==6.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
ruff==0.8.4
safety==2.3.5
